+++ flake.nix
@@ -19,11 +19,22 @@
       ];
       forAllSystems = nixpkgs.lib.genAttrs systems;
       pkgsFor = system: import nixpkgs { inherit system; };
+      pharosSource = nixpkgs.lib.cleanSourceWith {
+        src = ./.;
+        filter =
+          path: type:
+          let
+            relative = nixpkgs.lib.removePrefix "${toString ./.}/" (toString path);
+          in
+          nixpkgs.lib.cleanSourceFilter path type
+          && relative != "nix/tests"
+          && !nixpkgs.lib.hasPrefix "nix/tests/" relative;
+      };
       mkPharosPackage =
         pkgs: binaryName:
         pkgs.callPackage ./nix/packages/pharos.nix {
           inherit binaryName;
-          src = nixpkgs.lib.cleanSource ./.;
+          src = pharosSource;
         };
     in
     {
@@ -41,9 +52,20 @@
 
       checks = forAllSystems (
         system:
+        let
+          pkgs = pkgsFor system;
+        in
         {
           pharos-beacon = self.packages.${system}.pharos-beacon;
           pharosd = self.packages.${system}.pharosd;
+        }
+        // nixpkgs.lib.optionalAttrs (system == "x86_64-linux") {
+          pharos-beacon-vm = import ./nix/tests/pharos-beacon-vm.nix {
+            inherit pkgs;
+            pharosd = self.packages.${system}.pharosd;
+            pharosBeacon = self.packages.${system}.pharos-beacon;
+            pharosModule = ./nix/modules/pharos-beacon.nix;
+          };
         }
       );